pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "web-lab-scanner"
version = "0.1.0"
description = "Session-aware web application lab scanner for lab environments."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Ngoc Duy" }
]
dependencies = [
  "requests>=2.32.0",
  "beautifulsoup4>=4.12.0",
  "jinja2>=3.1.0",
  "pyyaml>=6.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0"
]

[project.scripts]
web-lab-scanner = "web_lab_scanner.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
